requirements.txt
Flask
requests
atlassian-python-api
openai
